pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 61"]

[project]
name = "blog"
authors = [
  { name="Mikhail Velikikh" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "mkdocs-material ~= 9.5.34",
    "mkdocs-rss-plugin ~= 1.15.0",
]
description = "Personal blog"
readme = "README.md"
requires-python = "==3.12"
version = "0.1"

[project.optional-dependencies]
dev = [
    "bs4",
]

[project.scripts]
convert-blogger-backup-to-md = "scripts.convert_blogger_backup_to_md:cli_main"
git-import = "scripts.git_import:cli_main"

[project.urls]
"Homepage" = "https://github.com/mvelikikh/mvelikikh.github.io"
